requirements.txt
streamlit>=1.28.0
beautifulsoup4>=4.12.2
google-genai>=0.2.0
openai>=1.0.0
requests>=2.31.0
protobuf>=5.29.5
